+++ packages/dapp/pages/create/index.tsx
@@ -1,69 +1,163 @@
-import { Button, Flex, Heading, Stack, Text } from '@chakra-ui/react';
-import { ChakraNextLink, useMediaStyles } from '@smartinvoicexyz/ui';
+import { Flex, Heading, Stack, Text } from '@chakra-ui/react';
+import {
+  ESCROW_STEPS,
+  INVOICE_TYPES,
+  TOASTS,
+} from '@smartinvoicexyz/constants';
+import {
+  EscrowDetailsForm,
+  FormConfirmation,
+  PaymentsForm,
+  ProjectDetailsForm,
+  RegisterSuccess,
+} from '@smartinvoicexyz/forms';
+import { useInvoiceCreate } from '@smartinvoicexyz/hooks';
+import {
+  Container,
+  NetworkChangeAlertModal,
+  StepInfo,
+  useMediaStyles,
+  useToast,
+} from '@smartinvoicexyz/ui';
+import { useQueryClient } from '@tanstack/react-query';
+import React, { useState } from 'react';
+import { useForm } from 'react-hook-form';
+import { Address, Hex } from 'viem';
 
-function SelectInvoiceType() {
-  const { primaryButtonSize } = useMediaStyles();
+import { useOverlay } from '../../contexts/OverlayContext';
+
+export function CreateInvoiceEscrow() {
+  const invoiceForm = useForm();
+  const toast = useToast();
+  const queryClient = useQueryClient();
+  const overlay = useOverlay();
+  const [currentStep, setCurrentStep] = useState<number>(1);
+  const [txHash, setTxHash] = useState<Hex>();
+
+  const [invoiceId, setInvoiceId] = useState<Address>();
+  const { headingSize, columnWidth } = useMediaStyles();
+
+  const nextStepHandler = () => {
+    setCurrentStep(currentStep + 1);
+  };
+
+  const goBackHandler = () => {
+    setCurrentStep(currentStep - 1);
+  };
+
+  const onTxSuccess = (result: Address) => {
+    toast.success(TOASTS.useInvoiceCreate.success);
+    // invalidate cache
+    queryClient.invalidateQueries({ queryKey: ['invoiceDetails'] });
+    queryClient.invalidateQueries({ queryKey: ['invoiceList'] });
+
+    setInvoiceId(result as Address);
+
+    // Send to Success step
+    nextStepHandler();
+  };
+
+  const { writeAsync, isLoading } = useInvoiceCreate({
+    invoiceForm,
+    toast,
+    onTxSuccess,
+  });
+
+  const handleSubmit = async () => {
+    const hash = await writeAsync?.();
+    setTxHash(hash);
+  };
 
   return (
-    <Flex direction="column" align="center" justify="center" gap={6}>
-      <Heading
-        fontWeight={700}
-        fontSize={50}
-        textAlign="center"
-        color="rgba(50, 60, 71, 1)"
-      >
-        Select an Invoice
-      </Heading>
-
-      <Text fontStyle="italic" color="grey">
-        Which type of invoice do you want to create?
-      </Text>
-
-      <Flex
-        direction={{ base: 'column', md: 'row' }}
-        columnGap={10}
-        rowGap={4}
-        width="100%"
-        align="stretch"
+    <Container overlay>
+      <Stack
+        direction={{ base: 'column', lg: 'column' }}
+        spacing="2rem"
+        align="center"
         justify="center"
-        paddingX={10}
+        w={columnWidth}
+        px="1rem"
+        my="2rem"
+        maxW="45rem"
       >
-        <ChakraNextLink href="/create/escrow">
-          <Button size={primaryButtonSize} variant="max">
-            <Stack spacing={6}>
-              <Heading>Escrow</Heading>
+        <NetworkChangeAlertModal {...overlay} />
 
-              <Stack fontSize={12} fontWeight="normal" textAlign="center">
-                <Text>Secure funds and release payments by milestones.</Text>
+        <Stack
+          spacing={{ base: '1.5rem', lg: '1rem' }}
+          w={{ base: '100%', md: 'auto' }}
+        >
+          <Heading fontWeight="700" fontSize={headingSize} textAlign="center">
+            Create an Escrow Invoice
+          </Heading>
 
-                <Text>Includes arbitration.</Text>
-              </Stack>
-              <Text fontSize={12} fontWeight="normal">
-                Recommended for medium to large projects
-              </Text>
-            </Stack>
-          </Button>
-        </ChakraNextLink>
+          <Text
+            color="#90A0B7"
+            as="i"
+            width="100%"
+            style={{ textIndent: 20 }}
+            align="center"
+          >
+            Note: All invoice data will be stored publicly on IPFS and can be
+            viewed by anyone. If you have privacy concerns, we recommend taking
+            care to add permissions to your project agreement document.
+          </Text>
 
-        <ChakraNextLink href="/create/instant">
-          <Button size={primaryButtonSize} variant="max">
-            <Stack spacing={6}>
-              <Heading>Instant</Heading>
+          <Flex
+            direction="column"
+            justify="space-between"
+            p="1rem"
+            bg="white"
+            borderRadius="0.5rem"
+            w="100%"
+          >
+            <StepInfo
+              stepNum={currentStep}
+              stepsDetails={ESCROW_STEPS}
+              goBack={goBackHandler}
+            />
+            {currentStep === 1 && (
+              <ProjectDetailsForm
+                invoiceForm={invoiceForm}
+                updateStep={nextStepHandler}
+                type={INVOICE_TYPES.Escrow}
+              />
+            )}
 
-              <Stack textAlign="center" fontSize={12} fontWeight="normal">
-                <Text wordBreak="break-word">Receive payment immediately.</Text>
+            {currentStep === 2 && (
+              <EscrowDetailsForm
+                invoiceForm={invoiceForm}
+                updateStep={nextStepHandler}
+              />
+            )}
 
-                <Text>Does NOT include arbitration.</Text>
-              </Stack>
-              <Text fontSize={12} fontWeight="normal">
-                Recommended for small projects
-              </Text>
-            </Stack>
-          </Button>
-        </ChakraNextLink>
-      </Flex>
-    </Flex>
+            {currentStep === 3 && (
+              <PaymentsForm
+                invoiceForm={invoiceForm}
+                updateStep={nextStepHandler}
+              />
+            )}
+
+            {currentStep === 4 && (
+              <FormConfirmation
+                invoiceForm={invoiceForm}
+                handleSubmit={handleSubmit}
+                canSubmit={!!writeAsync}
+                isLoading={isLoading}
+                type={INVOICE_TYPES.Escrow}
+              />
+            )}
+
+            {currentStep === 5 && (
+              <RegisterSuccess
+                invoiceId={invoiceId as Address}
+                txHash={txHash as Address}
+              />
+            )}
+          </Flex>
+        </Stack>
+      </Stack>
+    </Container>
   );
 }
 
-export default SelectInvoiceType;
+export default CreateInvoiceEscrow;
