+++ packages/ui/src/organisms/InvoiceDashboardTable.tsx
@@ -74,11 +74,6 @@ export function InvoiceDashboardTable() {
         // eslint-disable-next-line react/no-unstable-nested-components
         cell: info => <InvoiceDisplay cell={info} />,
       }),
-      /*
-      columnHelper.accessor('invoiceType', {
-        header: 'Type',
-        cell: info => _.capitalize(info.getValue()),
-      }),*/
       columnHelper.accessor(
         row => {
           const { provider, client, resolver } = row;
@@ -167,7 +162,7 @@ export function InvoiceDashboardTable() {
             </Heading>
           )}
 
-          <ChakraNextLink href="/create/escrow">
+          <ChakraNextLink href="/create">
             <Button size={primaryButtonSize} minW="250px" paddingY={6}>
               Create Invoice
             </Button>
@@ -190,7 +185,7 @@ export function InvoiceDashboardTable() {
             _hover={{ backgroundColor: 'rgba(61, 136, 248, 0.7)' }}
             _active={{ backgroundColor: 'rgba(61, 136, 248, 0.7)' }}
             color="white"
-            onClick={() => router.push('/create/escrow')}
+            onClick={() => router.push('/create')}
           >
             Create Invoice
           </Button>
